large refactor and additional features

+++ unclebudget/views.py
@@ -1,23 +1,151 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.core.exceptions import PermissionDenied
+from django.urls import reverse, reverse_lazy
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.views.generic import CreateView, DetailView, ListView, UpdateView
 
-from .models import Budget, Item, Transaction
+from unclebudget.models import Budget, Item, Transaction
 
-def budget_list(request):
-	budgets = Budget.objects.all().order_by('-last_modified')
-	return render(request, 'unclebudget/budget_list.html', locals())
+class BudgetCreate(LoginRequiredMixin, CreateView):
+	model = Budget
+	fields = ['name']
+	success_url = reverse_lazy('budget_list')
 
-def budget_detail(request, budget_id):
-	budget = Budget.objects.get(id=budget_id)
-	return render(request, 'unclebudget/budget_detail.html', locals())
+	def form_valid(self, form):
+		form.instance.user = self.request.user
+		return super().form_valid(form) 
 
-def item_detail(request, item_id):
-	item = Item.objects.get(id=item_id)
-	return render(request, 'unclebudget/item_detail.html', locals())
+class BudgetDetail(LoginRequiredMixin, DetailView):
+	model = Budget
+
+	def get_object(self):
+		budget = super().get_object()
+		if budget.user != self.request.user:
+			raise PermissionDenied
+		return budget
+
+class BudgetList(LoginRequiredMixin, ListView):
+	ordering = '-last_modified'
+
+	def get_queryset(self):
+		return Budget.objects.filter(user=self.request.user)
+
+class ExpenseCreate(LoginRequiredMixin, CreateView):
+	model = Item
+	fields = ['name', 'budgeted', 'singleton', 'transfer_to']
+
+	def form_valid(self, form):
+		budget = Budget.objects.get(id=self.kwargs['budget_pk'])
+		if budget.user != self.request.user:
+			raise PermissionDenied
+		form.instance.budget = budget
+		return super().form_valid(form) 
+	
+	def get_context_data(self):
+		context = super().get_context_data()
+		context['title'] = 'Add Expense Item'
+		return context
+
+	def get_success_url(self):
+		return reverse('budget_detail', kwargs={'pk': self.kwargs['budget_pk']})
+
+class ExpenseUpdate(LoginRequiredMixin, UpdateView):
+	model = Item
+	fields = ['name', 'budgeted']
+
+	def form_valid(self, form):
+		budget = Budget.objects.get(id=self.kwargs['budget_pk'])
+		if budget.user != self.request.user:
+			raise PermissionDenied
+		form.instance.budget = budget
+		return super().form_valid(form) 
+	
+	def get_context_data(self):
+		context = super().get_context_data()
+		context['title'] = 'Add Expense Item'
+		return context
+
+	def get_object(self):
+		item = super().get_object()
+		if item.budget.user != self.request.user:
+			raise PermissionDenied
+		return item
+
+	def get_success_url(self):
+		return reverse('budget_detail', kwargs={'pk': self.kwargs['budget_pk']})
+
+class IncomeCreate(LoginRequiredMixin, CreateView):
+	model = Item
+	fields = ['name', 'budgeted']
+
+	def form_valid(self, form):
+		budget = Budget.objects.get(id=self.kwargs['budget_pk'])
+		if budget.user != self.request.user:
+			raise PermissionDenied
+		form.instance.budget = budget
+		form.instance.income = True
+		return super().form_valid(form) 
+
+	def get_context_data(self):
+		context = super().get_context_data()
+		context['title'] = 'Add Income Item'
+		return context
+
+	def get_success_url(self):
+		return reverse('budget_detail', kwargs={'pk': self.kwargs['budget_pk']})
+
+class IncomeUpdate(LoginRequiredMixin, UpdateView):
+	model = Item
+	fields = ['name', 'budgeted']
+
+	def form_valid(self, form):
+		budget = Budget.objects.get(id=self.kwargs['budget_pk'])
+		if budget.user != self.request.user:
+			raise PermissionDenied
+		form.instance.budget = budget
+		form.instance.income = True
+		return super().form_valid(form) 
+
+	def get_context_data(self):
+		context = super().get_context_data()
+		context['title'] = 'Add Income Item'
+		return context
+
+	def get_object(self):
+		item = super().get_object()
+		if item.budget.user != self.request.user:
+			raise PermissionDenied
+		return item
+
+	def get_success_url(self):
+		return reverse('budget_detail', kwargs={'pk': self.kwargs['budget_pk']})
+
+class ItemDetail(LoginRequiredMixin, DetailView):
+	model = Item
+
+	def get_object(self):
+		item = super().get_object()
+		if item.budget.user != self.request.user:
+			raise PermissionDenied
+		return item
+
+def item_delete(request):
+	item = Item.objects.get(id=request.POST['itemID'])
+	if item.budget.user != request.user:
+		raise PermissionDenied
+
+	budgetID = item.budget.id
+	item.delete()
+	return redirect('budget_detail', budgetID)
 
 def transaction_create(request):
+	item = Item.objects.get(id=request.POST['itemID'])
+	if item.budget.user != request.user:
+		raise PermissionDenied
+
 	Transaction(
-		item=Item.objects.get(id=request.POST['itemID']),
+		item=item,
 		amount=request.POST['amount'],
 		comment=request.POST['comment']
 	).save()
@@ -25,6 +153,9 @@ def transaction_create(request):
 
 def transaction_delete(request):
 	transaction = Transaction.objects.get(id=request.POST['transactionID'])
+	if transaction.item.budget.user != request.user:
+		raise PermissionDenied
+
 	itemID = transaction.item.id
 	transaction.delete()
 	return redirect('item_detail', itemID)
